@@ -3,27 +3,23 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.10.0",
-    "@material-ui/icons": "^4.9.1",
     "axios": "^0.21.1",
-    "bootstrap": "^4.5.2",
-    "cors": "^2.8.5",
-    "jquery": "^3.5.1",
-    "npm": "^6.14.8",
-    "query-string": "^6.12.1",
-    "react": "^16.11.0",
-    "react-bootstrap": "^1.0.1",
-    "react-dom": "^16.11.0",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "^3.4.3",
-    "react-scroll-to-bottom": "^2.0.0",
-    "react-stomp": "^5.0.0",
-    "recharts": "^1.8.5",
-    "save": "^2.4.0",
-    "sockjs-client": "^1.4.0",
-    "stomp-client": "^0.9.0",
-    "stompjs": "^2.3.3",
-    "stringquery": "^1.0.8"
+    "bootstrap": "^5.3.2",
+    "date-fns": "^2.30.0",
+    "depcheck": "^1.4.3",
+    "jquery": "^3.7.1",
+    "popper.js": "^1.16.1",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.7.4",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.11.0",
+    "react-router-dom": "^6.13.0",
+    "react-scripts": "^5.0.1",
+    "react-stomp": "^5.1.0",
+    "recharts": "^2.7.1",
+    "sockjs-client": "^1.6.1",
+    "stompjs": "^2.3.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -46,5 +42,11 @@
       "last 1 safari version"
     ]
   },
-  "proxy": "http://localhost:8080"
+  "proxy": "http://localhost:8080",
+  "devDependencies": {
+    "eslint": "^8.50.0",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-unused-imports": "^3.0.0"
+  }
 }
