@@ -1,22 +1,21 @@
-import React, { useState } from 'react';
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import React, { useState } from "react";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 
 // Import custom styles for our application
-import './App.css';
+import "./App.css";
 
-import Auth from './services/Auth';
-import Navbar from './components/layout/Navbar';
+import Auth from "./services/Auth";
+import Navbar from "./components/home/navBar/Navbar";
 
 // Import pages
-import LoginPage from './components/auth/LoginPage';
-import HomePage from './components/home/HomePage';
-import PostsPage from './components/posts/PostsPage';
-import ProfilePage from './components/profile/ProfilePage';
-import Chat from './components/chat/Chat';
-import LoginFooter from './components/auth/LoginFooter';
+import Landing from "./components/landing/Landing";
+import Covid from "./components/home/covid/Covid";
+import Posts from "./components/home/posts/Posts";
+import Profile from "./components/home/profile/Profile";
+import News from "./components/home/news/News";
 //import ChatApp from "./ChatApp";
 
-import UserApi from './api/UserApi';
+import UserApi from "./api/UserApi";
 
 function App() {
   const [loggedIn, setLoggedIn] = useState(Auth.isLoggedIn());
@@ -25,39 +24,24 @@ function App() {
   UserApi.bindCurrentUserStateSetter(setUser);
 
   const loggedInRouter = (
-    <Router>
-      <Navbar onLogout={() => Auth.logout()} />
-
-      <div className="container mt-5">
-        <Switch>
-          <Route path="/posts">
-            <PostsPage />
-          </Route>
-          <Route path="/profile">
-            <ProfilePage user={user} />
-          </Route>
-
-          <Route path="/chat">
-            <Chat user={user} />
-          </Route>
-
-          <Route path="/">
-            <HomePage />
-          </Route>
-        </Switch>
+    <div className="bg-home">
+      <div>
+        <Router>
+          <Navbar onLogout={() => Auth.logout()} />
+          <div className="container mt-5">
+            <Routes>
+              <Route path="/posts" element={<Posts />} />
+              <Route path="/profile" element={<Profile user={user} />} />
+              <Route path="/news" element={<News />} />
+              <Route path="/" element={<Covid />} />
+            </Routes>
+          </div>
+        </Router>
       </div>
-      <LoginFooter />
-    </Router>
-  );
-
-  return loggedIn ? (
-    loggedInRouter
-  ) : (
-    <div>
-      <LoginPage />
-      <LoginFooter />
     </div>
   );
+
+  return loggedIn ? loggedInRouter : <Landing />;
 }
 
 export default App;
