@@ -1,25 +1,26 @@
-import Api from './Api';
+import Api from "./Api";
 
 class ChatApi {
   getAllChats() {
-    return Api.get('/chat');
+    return Api.get("/chat");
   }
 
   getChatById(id) {
-    return Api.get('/chat/' + id);
+    return Api.get("/chat/" + id);
   }
 
   createPChat(chat) {
-    return Api.post('/chat', chat);
+    return Api.post("/chat", chat);
   }
 
   updateChat(chat) {
-    return Api.put('/chat', chat);
+    return Api.put("/chat", chat);
   }
 
   deleteChat(id) {
-    return Api.delete('/chat/' + id);
+    return Api.delete("/chat/" + id);
   }
 }
 
-export default new ChatApi();
+const chatApiInstance = new ChatApi();
+export default chatApiInstance;
