@@ -1,31 +1,26 @@
 import Api from "./Api";
 
 class PostsApi {
+  getAllPosts() {
+    return Api.get("/posts");
+  }
 
-    
-    getAllPosts() {
-        
-        return Api.get('/posts');
-        
-    }
+  getPostById(id) {
+    return Api.get("/posts/" + id);
+  }
 
-    getPostById(id) {
-        return Api.get('/posts/'+id);
-    }
+  createPost(post) {
+    return Api.post("/posts", post);
+  }
 
-    createPost(post) {
-        return Api.post('/posts', post);
-    }
+  updatePost(post) {
+    return Api.put("/posts", post);
+  }
 
-    updatePost(post) {
-        return Api.put('/posts', post);
-    }
-
-    deletePost(id) {
-        return Api.delete('/posts/'+id);
-    } 
-    
+  deletePost(id) {
+    return Api.delete("/posts/" + id);
+  }
 }
 
-
-export default new PostsApi();
+const postApiInstance = new PostsApi();
+export default postApiInstance;
